day3_yolov8/requirements.txt
ultralytics==8.2.28
gradio==4.44.1
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.3.0
